+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2024-05-30
+
+- Added PagerDuty native support (#76).
+
+
 ## [0.8.0] - 2024-03-07
 
 - Fixes container silently running by exiting with non-zero status when configuration file is missing. (#70).